% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_prob.R
\name{equal_prob_wr}
\alias{equal_prob_wr}
\title{Equal Probability Sampling With Replacement}
\usage{
equal_prob_wr(N, n, method = "srs", nrep = 1L, prn = NULL, ...)
}
\arguments{
\item{N}{Population size (positive integer).}

\item{n}{Sample size (non-negative integer).}

\item{method}{The sampling method. Currently only \code{"srs"}.}

\item{nrep}{Number of replicate samples (default 1).}

\item{prn}{Optional vector of permanent random numbers for sample
coordination. Not currently used by any equal-probability WR method;
a warning is issued if provided.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An object of class \code{c("equal_prob", "wr", "sondage_sample")}.
When \code{nrep = 1}, \verb{$sample} is an integer vector and \verb{$hits} is an
integer vector. When \code{nrep > 1}, \verb{$sample} is a matrix (n x nrep)
and \verb{$hits} is a matrix (N x nrep).
}
\description{
Draws a simple random sample with replacement.
}
\examples{
set.seed(1)
s <- equal_prob_wr(10, 3)
s$sample
s$hits

}
\seealso{
\code{\link[=equal_prob_wor]{equal_prob_wor()}} for without-replacement designs,
\code{\link[=unequal_prob_wr]{unequal_prob_wr()}} for unequal probability designs.
}
