% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{is_registered_method}
\alias{is_registered_method}
\title{Check Whether a Method Is Registered}
\usage{
is_registered_method(name)
}
\arguments{
\item{name}{Method name (character string).}
}
\value{
\code{TRUE} if the method has been registered via
\code{\link[=register_method]{register_method()}}, \code{FALSE} otherwise.
}
\description{
Check Whether a Method Is Registered
}
\examples{
is_registered_method("foo")

}
\seealso{
\code{\link[=register_method]{register_method()}}
}
