% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.sondage_sample}
\alias{print.sondage_sample}
\alias{print.unequal_prob}
\alias{print.equal_prob}
\title{Print Sampling Design Objects}
\usage{
\method{print}{unequal_prob}(x, ...)

\method{print}{equal_prob}(x, ...)
}
\arguments{
\item{x}{A sampling design object of class \code{"sondage_sample"}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
\code{invisible(x)}.
}
\description{
Print Sampling Design Objects
}
