% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{registered_methods}
\alias{registered_methods}
\title{List Registered Custom Methods}
\usage{
registered_methods()
}
\value{
A character vector of registered method names (empty if none).
}
\description{
List Registered Custom Methods
}
\examples{
registered_methods()

}
\seealso{
\code{\link[=register_method]{register_method()}}
}
