% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{unregister_method}
\alias{unregister_method}
\title{Remove a Registered Method}
\usage{
unregister_method(name)
}
\arguments{
\item{name}{Method name to unregister.}
}
\value{
Invisible \code{TRUE} if the method was removed, \code{FALSE} if it
was not registered.
}
\description{
Remove a Registered Method
}
\examples{
unregister_method("nonexistent")

}
\seealso{
\code{\link[=register_method]{register_method()}}
}
