% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/up_joint.R
\name{up_brewer_joint}
\alias{up_brewer_joint}
\title{Joint Inclusion Probabilities using Brewer's Approximation}
\usage{
up_brewer_joint(pik)
}
\arguments{
\item{pik}{numeric vector of first-order inclusion probabilities.}
}
\value{
A symmetric N×N matrix of approximate joint probabilities.
}
\description{
Computes approximate joint inclusion probabilities using equation (18)
from Brewer & Donadio (2003).
}
\details{
Uses the approximation:
\deqn{c_i = \frac{n-1}{n - \frac{2n-1}{n-1}\pi_i + \frac{\sum \pi_k^2}{n-1}}}
\deqn{\tilde{\pi}_{ij} = \pi_i \pi_j \frac{c_i + c_j}{2}}

This has the best model-assisted properties among Brewer's approximations.
}
\examples{
pik <- c(0.2, 0.3, 0.5)
pikl <- up_brewer_joint(pik)

}
\references{
Brewer, K.R.W. and Donadio, M.E. (2003). The High Entropy Variance of the
Horvitz-Thompson Estimator. \emph{Survey Methodology}, 29(2), 189-196.
}
