% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sondage-package.R
\docType{package}
\name{sondage-package}
\alias{sondage}
\alias{sondage-package}
\title{sondage: Fast Survey Sampling Algorithms}
\description{
Fast implementations of survey sampling algorithms for single-stage probability sampling from finite populations. Provides equal probability methods (simple random sampling, systematic, Bernoulli), unequal probability methods (conditional Poisson / maximum entropy, Brewer, systematic PPS, Pareto piPS, sequential Poisson, Poisson, Chromy's minimum replacement, multinomial), and balanced sampling via the cube method. All sampling functions return design objects carrying sample indices, inclusion probabilities, and design metadata. Generics compute joint inclusion probabilities, pairwise expectations, and sampling covariances for variance estimation. Mostly based on algorithms from Tille (2006, \doi{10.1007/0-387-34240-0}).
}
\section{Unequal Probability Sampling}{

\itemize{
\item \code{\link[=unequal_prob_wor]{unequal_prob_wor()}} - Without replacement: CPS (maximum entropy),
Brewer, systematic PPS, Poisson, SPS (sequential Poisson),
Pareto
\item \code{\link[=unequal_prob_wr]{unequal_prob_wr()}} - With replacement: Chromy (minimum
replacement), multinomial PPS
}
}

\section{Equal Probability Sampling}{

\itemize{
\item \code{\link[=equal_prob_wor]{equal_prob_wor()}} - Without replacement: SRS, systematic,
Bernoulli (random size)
\item \code{\link[=equal_prob_wr]{equal_prob_wr()}} - With replacement: SRS
}
}

\section{Balanced Sampling}{

\itemize{
\item \code{\link[=balanced_wor]{balanced_wor()}} - Cube method (Deville & Tille, 2004) for
balanced sampling with unequal probabilities, with optional
stratification (Chauvet & Tille, 2006)
}
}

\section{Design Queries}{

All sampling functions return objects of class \code{"sondage_sample"}.
Use these generics to query the design:

\itemize{
\item \code{\link[=inclusion_prob]{inclusion_prob()}} - First-order inclusion probabilities
\item \code{\link[=expected_hits]{expected_hits()}} - Expected number of selections (WR)
\item \code{\link[=joint_inclusion_prob]{joint_inclusion_prob()}} - Joint inclusion probabilities (WOR)
\item \code{\link[=joint_expected_hits]{joint_expected_hits()}} - Pairwise expectations (WR)
\item \code{\link[=sampling_cov]{sampling_cov()}} - Sampling covariance matrix
}
}

\section{Utilities}{

\itemize{
\item \code{\link[=inclusion_prob]{inclusion_prob()}} - Compute inclusion probabilities from size
measures
\item \code{\link[=expected_hits]{expected_hits()}} - Compute expected hits from size measures
}
}

\references{
Tille, Y. (2006). \emph{Sampling Algorithms}. Springer Series in
Statistics.

Chromy, J.R. (2009). Some generalizations of the Horvitz-Thompson
estimator. \emph{Memorial JSM}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dickoa.gitlab.io/sondage}
  \item \url{https://gitlab.com/dickoa/sondage}
  \item Report bugs at \url{https://gitlab.com/dickoa/sondage/-/issues}
}

}
\author{
\strong{Maintainer}: Ahmadou Dicko \email{mail@ahmadoudicko.com} (\href{https://orcid.org/0000-0002-9654-7582}{ORCID})

Other contributors:
\itemize{
  \item Thomas Lumley \email{t.lumley@auckland.ac.nz} [contributor]
}

}
