% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sondage-package.R
\docType{package}
\name{sondage-package}
\alias{sondage}
\alias{sondage-package}
\title{sondage: Survey Sampling Algorithms}
\description{
Fast implementations of survey sampling algorithms for drawing samples
from finite populations. All functions return design objects that carry
sample indices, inclusion probabilities (or expected hits), and design
metadata.
}
\section{Unequal Probability Sampling}{

\itemize{
\item \code{\link[=unequal_prob_wor]{unequal_prob_wor()}} - Without replacement: CPS (maximum entropy),
Brewer, systematic PPS, Poisson, SPS (sequential Poisson),
Pareto
\item \code{\link[=unequal_prob_wr]{unequal_prob_wr()}} - With replacement: Chromy (minimum
replacement), multinomial PPS
}
}

\section{Equal Probability Sampling}{

\itemize{
\item \code{\link[=equal_prob_wor]{equal_prob_wor()}} - Without replacement: SRS, systematic,
Bernoulli (random size)
\item \code{\link[=equal_prob_wr]{equal_prob_wr()}} - With replacement: SRS
}
}

\section{Design Queries}{

All sampling functions return objects of class \code{"sondage_sample"}.
Use these generics to query the design:

\itemize{
\item \code{\link[=inclusion_prob]{inclusion_prob()}} - First-order inclusion probabilities
\item \code{\link[=expected_hits]{expected_hits()}} - Expected number of selections (WR)
\item \code{\link[=joint_inclusion_prob]{joint_inclusion_prob()}} - Joint inclusion probabilities (WOR)
\item \code{\link[=joint_expected_hits]{joint_expected_hits()}} - Pairwise expectations (WR)
\item \code{\link[=sampling_cov]{sampling_cov()}} - Sampling covariance matrix
}
}

\section{Utilities}{

\itemize{
\item \code{\link[=inclusion_prob]{inclusion_prob()}} - Compute inclusion probabilities from size
measures
\item \code{\link[=expected_hits]{expected_hits()}} - Compute expected hits from size measures
}
}

\references{
Tille, Y. (2006). \emph{Sampling Algorithms}. Springer Series in
Statistics.

Chromy, J.R. (2009). Some generalizations of the Horvitz-Thompson
estimator. \emph{Memorial JSM}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dickoa.gitlab.io/sondage}
  \item \url{https://gitlab.com/dickoa/sondage}
  \item Report bugs at \url{https://gitlab.com/dickoa/sondage/-/issues}
}

}
\author{
\strong{Maintainer}: Ahmadou Dicko \email{mail@ahmadoudicko.com} (\href{https://orcid.org/0000-0002-9654-7582}{ORCID})

}
