% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sondage-package.R
\docType{package}
\name{sondage-package}
\alias{sondage}
\alias{sondage-package}
\title{sondage: Survey Sampling Algorithms}
\description{
Fast implementations of survey sampling algorithms for drawing samples
from finite populations. All functions return indices for easy subsetting.
}
\section{Equal Probability Sampling}{

\itemize{
\item \code{\link[=srs]{srs()}} - Simple random sampling (with/without replacement)
\item \code{\link[=systematic]{systematic()}} - Systematic sampling
\item \code{\link[=bernoulli]{bernoulli()}} - Bernoulli sampling (random size)
}
}

\section{Unequal Probability Sampling}{

\itemize{
\item \code{\link[=up_maxent]{up_maxent()}} - Maximum entropy / Conditional Poisson Sampling
\item \code{\link[=up_brewer]{up_brewer()}} - Brewer's method
\item \code{\link[=up_systematic]{up_systematic()}} - Systematic PPS
\item \code{\link[=up_poisson]{up_poisson()}} - Poisson sampling (random size)
\item \code{\link[=up_multinomial]{up_multinomial()}} - PPS with replacement
\item \code{\link[=up_chromy]{up_chromy()}} - PPS with minimum replacement
}
}

\section{Utilities}{

\itemize{
\item \code{\link[=inclusion_prob]{inclusion_prob()}} - Compute inclusion probabilities from size measure
}
}

\references{
Tillé, Y. (2006). \emph{Sampling Algorithms}. Springer Series in Statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dickoa.gitlab.io/sondage}
  \item \url{https://gitlab.com/dickoa/sondage}
  \item Report bugs at \url{https://gitlab.com/dickoa/sondage/-/issues}
}

}
\author{
\strong{Maintainer}: Ahmadou Dicko \email{mail@ahmadoudicko.com} (\href{https://orcid.org/0000-0002-9654-7582}{ORCID})

}
